+++ src/ssj/help.c
@@ -43,7 +43,7 @@ help_print(const char* command_name)
 
 			" bt, backtrace    Show a list of all function calls currently on the stack     \n"
 			" bp, breakpoint   Set a breakpoint at file:line (e.g. scripts/eaty-pig.js:812) \n"
-			" cb, clearbreak   Clear a breakpoint set at file:line (see 'breakpoint')       \n"
+			" cb, clear        Clear a breakpoint set at file:line (see 'breakpoint')       \n"
 			" c,  continue     Run either until a breakpoint is hit or an error is thrown   \n"
 			" d,  down         Move down the call stack (inwards) from the selected frame   \n"
 			" e,  eval         Evaluate a JavaScript expression                             \n"
@@ -78,13 +78,13 @@ help_print(const char* command_name)
 			"invaluable.  When the breakpoint is reached, execution will pause, leaving SSj \n"
 			"in control.                                                                    \n\n"
 			"The ID number of the set breakpoint will be given, which will be needed if you \n"
-			"want to clear it later using 'clearbreak'.                                        \n\n"
+			"want to clear it later using 'clear'.                                          \n\n"
 			"SYNTAX:                                                                        \n"
 			"    breakpoint             - to list active breakpoints                        \n"
 			"    breakpoint <file:line> - to set a breakpoint                               \n"
 		);
 	}
-	else if (strcmp(command_name, "clearbreak") == 0) {
+	else if (strcmp(command_name, "clear") == 0) {
 		printf(
 			"Clear the breakpoint <index> established using 'breakpoint'.  When you no      \n"
 			"longer need a breakpoint, you should clear to prevent execution from pausing in\n"
@@ -92,7 +92,7 @@ help_print(const char* command_name)
 			"To list active breakpoints and their indices, simply enter 'breakpoint' with no\n"
 			"arguments.                                                                     \n\n"
 			"SYNTAX:                                                                        \n"
-			"    clearbreak <index>                                                            \n"
+			"    clear <index>                                                              \n"
 		);
 	}
 	else if (strcmp(command_name, "continue") == 0) {