+++ src/ssj/session.c
@@ -73,7 +73,7 @@ command_db[] =
 {
 	"backtrace",  "bt", "",
 	"breakpoint", "bp", "~f",
-	"clearbreak", "cb", "n",
+	"clear",      "cb", "n",
 	"continue",   "c",  "",
 	"down",       "d",  "~n",
 	"eval",       "e",  "*",
@@ -90,55 +90,55 @@ command_db[] =
 	"help",       "h",  "~s",
 };
 
-static void        autoselect_frame  (session_t* obj);
-static void        clear_auto_action (session_t* obj);
-static void        do_command_line   (session_t* obj);
+static void        autoselect_frame  (session_t* session);
+static void        clear_auto_action (session_t* session);
+static void        do_command_line   (session_t* session);
 static const char* find_verb         (const char* abbrev, const char* *o_pattern);
 static const char* resolve_command   (command_t* cmd);
-static void        handle_backtrace  (session_t* obj, command_t* cmd);
-static void        handle_breakpoint (session_t* obj, command_t* cmd);
-static void        handle_clearbreak (session_t* obj, command_t* cmd);
-static void        handle_eval       (session_t* obj, command_t* cmd, bool is_verbose);
-static void        handle_frame      (session_t* obj, command_t* cmd);
-static void        handle_help       (session_t* obj, command_t* cmd);
-static void        handle_list       (session_t* obj, command_t* cmd);
-static void        handle_resume     (session_t* obj, command_t* cmd, resume_op_t op);
-static void        handle_up_down    (session_t* obj, command_t* cmd, int direction);
-static void        handle_vars       (session_t* obj, command_t* cmd);
-static void        handle_where      (session_t* obj, command_t* cmd);
-static void        handle_quit       (session_t* obj, command_t* cmd);
-static void        preview_frame     (session_t* obj, int frame);
-static bool        validate_args     (const command_t* this, const char* verb_name, const char* pattern);
+static void        handle_backtrace  (session_t* session, command_t* cmd);
+static void        handle_breakpoint (session_t* session, command_t* cmd);
+static void        handle_clear      (session_t* session, command_t* cmd);
+static void        handle_eval       (session_t* session, command_t* cmd, bool is_verbose);
+static void        handle_frame      (session_t* session, command_t* cmd);
+static void        handle_help       (session_t* session, command_t* cmd);
+static void        handle_list       (session_t* session, command_t* cmd);
+static void        handle_resume     (session_t* session, command_t* cmd, resume_op_t op);
+static void        handle_up_down    (session_t* session, command_t* cmd, int direction);
+static void        handle_vars       (session_t* session, command_t* cmd);
+static void        handle_where      (session_t* session, command_t* cmd);
+static void        handle_quit       (session_t* session, command_t* cmd);
+static void        preview_frame     (session_t* session, int frame);
+static bool        validate_args     (const command_t* command, const char* verb_name, const char* pattern);
 
 session_t*
 session_new(inferior_t* inferior)
 {
-	session_t* obj;
+	session_t* session;
 
-	obj = calloc(1, sizeof(session_t));
-	obj->inferior = inferior;
-	return obj;
+	session = calloc(1, sizeof(session_t));
+	session->inferior = inferior;
+	return session;
 }
 
 void
-session_free(session_t* obj)
+session_free(session_t* it)
 {
-	free(obj);
+	free(it);
 }
 
 void
-session_run(session_t* obj, bool run_now)
+session_run(session_t* it, bool run_now)
 {
 	if (run_now) {
-		inferior_resume(obj->inferior, OP_RESUME);
+		inferior_resume(it->inferior, OP_RESUME);
 	}
-	if (inferior_attached(obj->inferior)) {
-		autoselect_frame(obj);
-		preview_frame(obj, obj->frame);
+	if (inferior_attached(it->inferior)) {
+		autoselect_frame(it);
+		preview_frame(it, it->frame);
 	}
 
-	while (inferior_attached(obj->inferior))
-		do_command_line(obj);
+	while (inferior_attached(it->inferior))
+		do_command_line(it);
 	printf("the SSj debugger has been detached.\n");
 }
 
@@ -175,9 +175,9 @@ find_verb(const char* abbrev, const char* *out_pattern)
 	if (num_matches == 1)
 		return matches[0];
 	else if (num_matches > 1) {
-		printf("'%s': abbreviated name is ambiguous between:\n", abbrev);
+		printf("command '%s' is ambiguous.  did you want:\n", abbrev);
 		for (i = 0; i < num_matches; ++i)
-			printf("    * %s\n", matches[i]);
+			printf("  - %s\n", matches[i]);
 		return NULL;
 	}
 	else {
@@ -187,29 +187,29 @@ find_verb(const char* abbrev, const char* *out_pattern)
 }
 
 static void
-autoselect_frame(session_t* obj)
+autoselect_frame(session_t* session)
 {
 	const backtrace_t* calls;
 
-	calls = inferior_get_calls(obj->inferior);
-	obj->frame = 0;
-	while (backtrace_get_linenum(calls, obj->frame) == 0)
-		++obj->frame;
+	calls = inferior_get_calls(session->inferior);
+	session->frame = 0;
+	while (backtrace_get_linenum(calls, session->frame) == 0)
+		++session->frame;
 }
 
 static void
-clear_auto_action(session_t* obj)
+clear_auto_action(session_t* session)
 {
-	switch (obj->auto_action) {
+	switch (session->auto_action) {
 	case AUTO_LIST:
-		free(obj->list_filename);
+		free(session->list_filename);
 		break;
 	}
-	obj->auto_action = AUTO_NONE;
+	session->auto_action = AUTO_NONE;
 }
 
 static void
-do_command_line(session_t* obj)
+do_command_line(session_t* session)
 {
 	char               buffer[4096];
 	const backtrace_t* calls;
@@ -223,10 +223,10 @@ do_command_line(session_t* obj)
 
 	int idx;
 
-	calls = inferior_get_calls(obj->inferior);
-	function_name = backtrace_get_call_name(calls, obj->frame);
-	filename = backtrace_get_filename(calls, obj->frame);
-	line_no = backtrace_get_linenum(calls, obj->frame);
+	calls = inferior_get_calls(session->inferior);
+	function_name = backtrace_get_call_name(calls, session->frame);
+	filename = backtrace_get_filename(calls, session->frame);
+	line_no = backtrace_get_linenum(calls, session->frame);
 	if (line_no != 0)
 		printf("\n\33[36;1m%s:%d %s\33[m\n\33[33;1m(ssj)\33[m ", filename, line_no, function_name);
 	else
@@ -247,16 +247,16 @@ do_command_line(session_t* obj)
 		goto finished;
 
 	// if the command line is empty, this is a cue from the user that we should
-	// repeat the last command.  the implementation of this is a bit hacky and would benefit
+	// repeat the last command.  the implementation of this is very hacky and would benefit
 	// from some refactoring in the future.
 	if (command_len(command) == 0) {
 		command_free(command);
-		switch (obj->auto_action) {
+		switch (session->auto_action) {
 		case AUTO_CONTINUE:
 			command = command_parse("continue");
 			break;
 		case AUTO_LIST:
-			synth = strnewf("list %d \"%s\":%d", obj->list_num_lines, obj->list_filename, obj->list_linenum);
+			synth = strnewf("list %d \"%s\":%d", session->list_num_lines, session->list_filename, session->list_linenum);
 			command = command_parse(synth);
 			free(synth);
 			break;
@@ -270,9 +270,9 @@ do_command_line(session_t* obj)
 			command = command_parse("stepover");
 			break;
 		case AUTO_UP_DOWN:
-			if (obj->up_down_direction > 0)
+			if (session->up_down_direction > 0)
 				command = command_parse("up");
-			else if (obj->up_down_direction < 0)
+			else if (session->up_down_direction < 0)
 				command = command_parse("down");
 			else  // crash prevention
 				command = command_parse("frame");
@@ -291,41 +291,41 @@ do_command_line(session_t* obj)
 	// figure out which handler to run based on the command name. this could
 	// probably be refactored to get rid of the massive if/elseif tower, but for
 	// now it serves its purpose.
-	clear_auto_action(obj);
+	clear_auto_action(session);
 	if (strcmp(verb, "quit") == 0)
-		handle_quit(obj, command);
+		handle_quit(session, command);
 	else if (strcmp(verb, "help") == 0)
-		handle_help(obj, command);
+		handle_help(session, command);
 	else if (strcmp(verb, "backtrace") == 0)
-		handle_backtrace(obj, command);
+		handle_backtrace(session, command);
 	else if (strcmp(verb, "breakpoint") == 0)
-		handle_breakpoint(obj, command);
+		handle_breakpoint(session, command);
 	else if (strcmp(verb, "up") == 0)
-		handle_up_down(obj, command, +1);
+		handle_up_down(session, command, +1);
 	else if (strcmp(verb, "down") == 0)
-		handle_up_down(obj, command, -1);
-	else if (strcmp(verb, "clearbreak") == 0)
-		handle_clearbreak(obj, command);
+		handle_up_down(session, command, -1);
+	else if (strcmp(verb, "clear") == 0)
+		handle_clear(session, command);
 	else if (strcmp(verb, "continue") == 0)
-		handle_resume(obj, command, OP_RESUME);
+		handle_resume(session, command, OP_RESUME);
 	else if (strcmp(verb, "eval") == 0)
-		handle_eval(obj, command, false);
+		handle_eval(session, command, false);
 	else if (strcmp(verb, "examine") == 0)
-		handle_eval(obj, command, true);
+		handle_eval(session, command, true);
 	else if (strcmp(verb, "frame") == 0)
-		handle_frame(obj, command);
+		handle_frame(session, command);
 	else if (strcmp(verb, "list") == 0)
-		handle_list(obj, command);
+		handle_list(session, command);
 	else if (strcmp(verb, "stepover") == 0)
-		handle_resume(obj, command, OP_STEP_OVER);
+		handle_resume(session, command, OP_STEP_OVER);
 	else if (strcmp(verb, "stepin") == 0)
-		handle_resume(obj, command, OP_STEP_IN);
+		handle_resume(session, command, OP_STEP_IN);
 	else if (strcmp(verb, "stepout") == 0)
-		handle_resume(obj, command, OP_STEP_OUT);
+		handle_resume(session, command, OP_STEP_OUT);
 	else if (strcmp(verb, "vars") == 0)
-		handle_vars(obj, command);
+		handle_vars(session, command);
 	else if (strcmp(verb, "where") == 0)
-		handle_where(obj, command);
+		handle_where(session, command);
 	else
 		printf("'%s': not implemented.\n", verb);
 
@@ -349,17 +349,17 @@ resolve_command(command_t* command)
 }
 
 static void
-handle_backtrace(session_t* obj, command_t* cmd)
+handle_backtrace(session_t* session, command_t* cmd)
 {
 	const backtrace_t* calls;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
-	backtrace_print(calls, obj->frame, true);
+	backtrace_print(calls, session->frame, true);
 }
 
 static void
-handle_breakpoint(session_t* obj, command_t* cmd)
+handle_breakpoint(session_t* session, command_t* cmd)
 {
 	const char*      filename;
 	int              handle;
@@ -369,38 +369,38 @@ handle_breakpoint(session_t* obj, command_t* cmd)
 	int i;
 
 	if (command_len(cmd) < 2) {
-		if (obj->num_breaks <= 0)
+		if (session->num_breaks <= 0)
 			printf("no breakpoints are currently set.\n");
 		else {
-			for (i = 0; i < obj->num_breaks; ++i) {
+			for (i = 0; i < session->num_breaks; ++i) {
 				printf("#%2d: breakpoint at %s:%d\n", i,
-					obj->breaks[i].filename,
-					obj->breaks[i].linenum);
+					session->breaks[i].filename,
+					session->breaks[i].linenum);
 			}
 		}
 	}
 	else {
 		filename = command_get_string(cmd, 1);
 		linenum = command_get_int(cmd, 1);
-		if ((handle = inferior_add_breakpoint(obj->inferior, filename, linenum)) < 0)
+		if ((handle = inferior_add_breakpoint(session->inferior, filename, linenum)) < 0)
 			printf("SSj was unable to set the breakpoint.\n");
 		else {
-			i = obj->num_breaks++;
-			obj->breaks = realloc(obj->breaks, obj->num_breaks * sizeof(struct breakpoint));
-			obj->breaks[i].handle = handle;
-			obj->breaks[i].filename = strdup(command_get_string(cmd, 1));
-			obj->breaks[i].linenum = command_get_int(cmd, 1);
+			i = session->num_breaks++;
+			session->breaks = realloc(session->breaks, session->num_breaks * sizeof(struct breakpoint));
+			session->breaks[i].handle = handle;
+			session->breaks[i].filename = strdup(command_get_string(cmd, 1));
+			session->breaks[i].linenum = command_get_int(cmd, 1);
 			printf("breakpoint #%2d set at %s:%d.\n", i,
-				obj->breaks[i].filename,
-				obj->breaks[i].linenum);
-			if ((listing = inferior_get_listing(obj->inferior, filename)))
+				session->breaks[i].filename,
+				session->breaks[i].linenum);
+			if ((listing = inferior_get_listing(session->inferior, filename)))
 				listing_print(listing, linenum, 1, 0);
 		}
 	}
 }
 
 static void
-handle_clearbreak(session_t* obj, command_t* cmd)
+handle_clear(session_t* session, command_t* cmd)
 {
 	const char*      filename;
 	int              handle;
@@ -409,50 +409,52 @@ handle_clearbreak(session_t* obj, command_t* cmd)
 	const listing_t* listing;
 
 	index = command_get_int(cmd, 1);
-	if (obj->num_breaks <= 0)
-		printf("no breakpoints to clear.\n");
-	else if (index > obj->num_breaks)
-		printf("invalid breakpoint index, valid range is 0-%d.", obj->num_breaks);
+	if (session->num_breaks <= 0) {
+		printf("there are no breakpoints to clear.\n");
+	}
+	else if (index < 0 || index >= session->num_breaks) {
+		printf("invalid breakpoint index, valid range is [0,%d].", session->num_breaks - 1);
+	}
 	else {
-		handle = obj->breaks[index].handle;
-		filename = obj->breaks[index].filename;
-		linenum = obj->breaks[index].linenum;
-		if (!inferior_clear_breakpoint(obj->inferior, handle))
+		handle = session->breaks[index].handle;
+		filename = session->breaks[index].filename;
+		linenum = session->breaks[index].linenum;
+		if (!inferior_clear_breakpoint(session->inferior, handle))
 			return;
 		printf("cleared breakpoint #%2d at %s:%d.\n", index,
-			obj->breaks[index].filename, obj->breaks[index].linenum);
-		if ((listing = inferior_get_listing(obj->inferior, filename)))
+			session->breaks[index].filename, session->breaks[index].linenum);
+		if ((listing = inferior_get_listing(session->inferior, filename)))
 			listing_print(listing, linenum, 1, 0);
 	}
 }
 
 static void
-handle_resume(session_t* obj, command_t* cmd, resume_op_t op)
+handle_resume(session_t* session, command_t* cmd, resume_op_t op)
 {
 	switch (op) {
 	case OP_RESUME:
-		obj->auto_action = AUTO_CONTINUE;
+		session->auto_action = AUTO_CONTINUE;
 		break;
 	case OP_STEP_IN:
-		obj->auto_action = AUTO_STEP_IN;
+		session->auto_action = AUTO_STEP_IN;
 		break;
 	case OP_STEP_OUT:
-		obj->auto_action = AUTO_STEP_OUT;
+		session->auto_action = AUTO_STEP_OUT;
 		break;
 	case OP_STEP_OVER:
-		obj->auto_action = AUTO_STEP_OVER;
+		session->auto_action = AUTO_STEP_OVER;
 		break;
 	}
 
-	inferior_resume(obj->inferior, op);
-	if (inferior_attached(obj->inferior)) {
-		autoselect_frame(obj);
-		preview_frame(obj, obj->frame);
+	inferior_resume(session->inferior, op);
+	if (inferior_attached(session->inferior)) {
+		autoselect_frame(session);
+		preview_frame(session, session->frame);
 	}
 }
 
 static void
-handle_eval(session_t* obj, command_t* cmd, bool verbose)
+handle_eval(session_t* session, command_t* cmd, bool verbose)
 {
 	const char*      expr = NULL;
 	const ki_atom_t* getter;
@@ -473,7 +475,7 @@ handle_eval(session_t* obj, command_t* cmd, bool verbose)
 	}
 	else {
 		expr = command_get_rest(cmd, 1);
-		result = inferior_eval(obj->inferior, expr, obj->frame, &is_error);
+		result = inferior_eval(session->inferior, expr, session->frame, &is_error);
 		handle = ki_atom_handle(result);
 		if (ki_atom_type(result) != KI_REF) {  // primitive value?
 			if (!is_error) {
@@ -488,7 +490,7 @@ handle_eval(session_t* obj, command_t* cmd, bool verbose)
 		}
 	}
 
-	if (!(object = inferior_get_object(obj->inferior, handle, verbose)))
+	if (!(object = inferior_get_object(session->inferior, handle, verbose)))
 		return;
 	if (is_error)
 		printf("\33[31;1m");
@@ -531,26 +533,26 @@ handle_eval(session_t* obj, command_t* cmd, bool verbose)
 }
 
 static void
-handle_frame(session_t* obj, command_t* cmd)
+handle_frame(session_t* session, command_t* cmd)
 {
 	const backtrace_t* calls;
 	int                frame;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
-	frame = obj->frame;
+	frame = session->frame;
 	if (command_len(cmd) >= 2)
 		frame = command_get_int(cmd, 1);
 	if (frame < 0 || frame >= backtrace_len(calls))
 		printf("stack frame #%2d doesn't exist.\n", frame);
 	else {
-		obj->frame = frame;
-		preview_frame(obj, obj->frame);
+		session->frame = frame;
+		preview_frame(session, session->frame);
 	}
 }
 
 static void
-handle_help(session_t* obj, command_t* cmd)
+handle_help(session_t* session, command_t* cmd)
 {
 	const char* verb;
 
@@ -564,68 +566,72 @@ handle_help(session_t* obj, command_t* cmd)
 }
 
 static void
-handle_list(session_t* obj, command_t* cmd)
+handle_list(session_t* session, command_t* cmd)
 {
 	const char*        active_filename;
 	int                active_lineno = 0;
 	const backtrace_t* calls;
 	const char*        filename;
-	int                lineno;
+	int                line_number;
 	const listing_t*   listing;
 	int                num_lines = 10;
 
-	calls = inferior_get_calls(obj->inferior);
-	active_filename = backtrace_get_filename(calls, obj->frame);
-	active_lineno = backtrace_get_linenum(calls, obj->frame);
+	calls = inferior_get_calls(session->inferior);
+	active_filename = backtrace_get_filename(calls, session->frame);
+	active_lineno = backtrace_get_linenum(calls, session->frame);
 	filename = active_filename;
-	lineno = active_lineno;
+	line_number = active_lineno;
 	if (command_len(cmd) >= 2)
 		num_lines = command_get_int(cmd, 1);
 	if (command_len(cmd) >= 3) {
 		filename = command_get_string(cmd, 2);
-		lineno = command_get_int(cmd, 2);
+		line_number = command_get_int(cmd, 2);
 	}
-	if (!(listing = inferior_get_listing(obj->inferior, filename)))
-		printf("source unavailable for %s.\n", filename);
+	if (!(listing = inferior_get_listing(session->inferior, filename))) {
+		printf("no source code is available for '%s'.\n", filename);
+	}
 	else {
 		if (strcmp(filename, active_filename) != 0)
 			active_lineno = 0;
-		listing_print(listing, lineno, num_lines, active_lineno);
-		obj->list_num_lines = num_lines;
-		obj->list_filename = strdup(filename);
-		obj->list_linenum = lineno + num_lines;
-		obj->auto_action = AUTO_LIST;
+		line_number -= num_lines / 2;
+		if (line_number < 1)
+			line_number = 1;
+		listing_print(listing, line_number, num_lines, active_lineno);
+		session->list_num_lines = num_lines;
+		session->list_filename = strdup(filename);
+		session->list_linenum = (line_number + num_lines / 2) + num_lines;
+		session->auto_action = AUTO_LIST;
 	}
 }
 
 static void
-handle_up_down(session_t* obj, command_t* cmd, int direction)
+handle_up_down(session_t* session, command_t* cmd, int direction)
 {
 	const backtrace_t* calls;
 	int                new_frame;
 	int                num_steps;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
-	new_frame = obj->frame + direction;
+	new_frame = session->frame + direction;
 	if (new_frame >= backtrace_len(calls))
 		printf("innermost frame: can't go up any further.\n");
 	else if (new_frame < 0)
 		printf("outermost frame: can't go down any further.\n");
 	else {
 		num_steps = command_len(cmd) >= 2 ? command_get_int(cmd, 1) : 1;
-		new_frame = obj->frame + num_steps * direction;
-		obj->frame = new_frame < 0 ? 0
+		new_frame = session->frame + num_steps * direction;
+		session->frame = new_frame < 0 ? 0
 			: new_frame >= backtrace_len(calls) ? backtrace_len(calls) - 1
 			: new_frame;
-		preview_frame(obj, obj->frame);
+		preview_frame(session, session->frame);
 	}
-	obj->auto_action = AUTO_UP_DOWN;
-	obj->up_down_direction = direction;
+	session->auto_action = AUTO_UP_DOWN;
+	session->up_down_direction = direction;
 }
 
 static void
-handle_vars(session_t* obj, command_t* cmd)
+handle_vars(session_t* session, command_t* cmd)
 {
 	const backtrace_t* calls;
 	const char*        call_name;
@@ -637,9 +643,9 @@ handle_vars(session_t* obj, command_t* cmd)
 
 	int i;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
-	if (!(vars = inferior_get_vars(obj->inferior, obj->frame)))
+	if (!(vars = inferior_get_vars(session->inferior, session->frame)))
 		return;
 	for (i = 0; i < objview_len(vars); ++i) {
 		var_name = objview_get_key(vars, i);
@@ -647,7 +653,7 @@ handle_vars(session_t* obj, command_t* cmd)
 			max_len = (int)strlen(var_name);
 	}
 	if (objview_len(vars) == 0) {
-		call_name = backtrace_get_call_name(calls, obj->frame);
+		call_name = backtrace_get_call_name(calls, session->frame);
 		printf("no local variables in scope for %s\n", call_name);
 	}
 	for (i = 0; i < objview_len(vars); ++i) {
@@ -661,33 +667,33 @@ handle_vars(session_t* obj, command_t* cmd)
 }
 
 static void
-handle_where(session_t* obj, command_t* cmd)
+handle_where(session_t* session, command_t* cmd)
 {
 	int                frame = 0;
 	const backtrace_t* calls;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
 	while (backtrace_get_linenum(calls, frame) <= 0)
 		++frame;
-	preview_frame(obj, frame);
+	preview_frame(session, frame);
 }
 
 static void
-handle_quit(session_t* obj, command_t* cmd)
+handle_quit(session_t* session, command_t* cmd)
 {
-	inferior_detach(obj->inferior);
+	inferior_detach(session->inferior);
 }
 
 static void
-preview_frame(session_t* obj, int frame)
+preview_frame(session_t* session, int frame)
 {
 	const backtrace_t* calls;
 	const char*        filename;
 	int                lineno;
 	const listing_t*   listing;
 
-	if (!(calls = inferior_get_calls(obj->inferior)))
+	if (!(calls = inferior_get_calls(session->inferior)))
 		return;
 	backtrace_print(calls, frame, false);
 	filename = backtrace_get_filename(calls, frame);
@@ -695,7 +701,7 @@ preview_frame(session_t* obj, int frame)
 	if (lineno == 0)
 		printf("system call, no source provided.\n");
 	else {
-		if (!(listing = inferior_get_listing(obj->inferior, filename)))
+		if (!(listing = inferior_get_listing(session->inferior, filename)))
 			printf("source unavailable for %s.\n", filename);
 		else
 			listing_print(listing, lineno, 1, lineno);
@@ -703,7 +709,7 @@ preview_frame(session_t* obj, int frame)
 }
 
 static bool
-validate_args(const command_t* this, const char* verb_name, const char* pattern)
+validate_args(const command_t* command, const char* verb_name, const char* pattern)
 {
 	int         index = 0;
 	int         want_num_args;
@@ -715,12 +721,12 @@ validate_args(const command_t* this, const char* verb_name, const char* pattern)
 		want_num_args = (int)(strchr(pattern, '~') - pattern);
 	else
 		want_num_args = (int)strlen(pattern);
-	if (command_len(this) - 1 < want_num_args) {
+	if (command_len(command) - 1 < want_num_args) {
 		printf("'%s': expected at least %d arguments.\n", verb_name, want_num_args);
 		return false;
 	}
 	p_type = pattern;
-	while (index < command_len(this) - 1) {
+	while (index < command_len(command) - 1) {
 		if (*p_type == '~')
 			++p_type;
 		if (*p_type == '\0')
@@ -744,7 +750,7 @@ validate_args(const command_t* this, const char* verb_name, const char* pattern)
 			want_type = "string";
 			break;
 		}
-		if (want_tag != TOK_ANY && command_get_tag(this, index + 1) != want_tag)
+		if (want_tag != TOK_ANY && command_get_tag(command, index + 1) != want_tag)
 			goto wrong_type;
 		++p_type;
 		++index;